PROX-393: rename classes, folder
