+++ src/main/java/com/rbkmoney/proxy/mocketbank/utils/reader/MobilePhoneReader.java
@@ -1,22 +1,22 @@
-package com.rbkmoney.proxy.mocketbank.utils.extractor;
+package com.rbkmoney.proxy.mocketbank.utils.reader;
 
-import com.rbkmoney.proxy.mocketbank.utils.model.Card;
+import com.rbkmoney.proxy.mocketbank.utils.mobilephone.MobilePhone;
 import org.springframework.stereotype.Component;
 
 import java.io.InputStream;
 import java.util.List;
 
 @Component
-public class CardReader implements BeanReader<Card> {
+public class MobilePhoneReader implements BeanReader<MobilePhone> {
 
     private static final String REGEXP = ", ";
 
     @Override
-    public List<Card> readList(InputStream is) {
+    public List<MobilePhone> readList(InputStream is) {
         return extractListFromFile(is,
                 line -> {
                     String[] p = line.split(REGEXP);
-                    return new Card(p[0], p[1], p[2]);
+                    return new MobilePhone(p[0], p[1], p[2]);
                 });
     }
 }
