+++ src/main/java/com/rbkmoney/proxy/mocketbank/validator/WithdrawalValidator.java
@@ -2,7 +2,7 @@ package com.rbkmoney.proxy.mocketbank.validator;
 
 import com.rbkmoney.damsel.withdrawals.provider_adapter.GetQuoteParams;
 import com.rbkmoney.proxy.mocketbank.exception.WithdrawalException;
-import com.rbkmoney.proxy.mocketbank.handler.oct.converter.CurrencyConverter;
+import com.rbkmoney.proxy.mocketbank.service.oct.verification.CurrencyVerification;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 import org.springframework.stereotype.Component;
@@ -15,13 +15,13 @@ public class WithdrawalValidator implements Validator<GetQuoteParams> {
 
     public void validate(GetQuoteParams getQuoteParams, Map<String, String> options) {
         validateRequiredFields(options);
-        if (CurrencyConverter.isCryptoCurrency(getQuoteParams.getCurrencyFrom()) && CurrencyConverter.isCryptoCurrency(getQuoteParams.getCurrencyTo())) {
+        if (CurrencyVerification.isCryptoCurrency(getQuoteParams.getCurrencyFrom()) && CurrencyVerification.isCryptoCurrency(getQuoteParams.getCurrencyTo())) {
             throw new WithdrawalException("Can't exchange crypto currency to crypto currency");
         }
-        if (CurrencyConverter.isCurrencyEquals(getQuoteParams.getCurrencyFrom(), getQuoteParams.getCurrencyTo())) {
+        if (CurrencyVerification.isCurrencyEquals(getQuoteParams.getCurrencyFrom(), getQuoteParams.getCurrencyTo())) {
             throw new WithdrawalException("Can't exchange equals currency");
         }
-        if (!CurrencyConverter.isExchangeCurrencyEqualsOtherCurrency(getQuoteParams)) {
+        if (!CurrencyVerification.isExchangeCurrencyEqualsOtherCurrency(getQuoteParams)) {
             throw new WithdrawalException("Can't exchange equals currency, all currency is different");
         }
     }
